Refactoring and Documentation done.
